Update cost-per-use fields to be decimals.

+++ ivetl/models/subscription_cost_per_use_delta.py
@@ -8,9 +8,9 @@ class SubscriptionCostPerUseByBundleStatDelta(Model):
     bundle_name = columns.Text(primary_key=True)
     usage_date = columns.DateTime(primary_key=True)
     time_slice = columns.Text(primary_key=True)
-    previous_cost_per_use = columns.Float()
-    current_cost_per_use = columns.Float()
-    absolute_delta = columns.Integer()
+    previous_cost_per_use = columns.Decimal()
+    current_cost_per_use = columns.Decimal()
+    absolute_delta = columns.Decimal()
     percentage_delta = columns.Float()
 
     __table_name__ = 'subscription_cpu_by_bundle_stat_delta'
@@ -21,9 +21,9 @@ class SubscriptionCostPerUseBySubscriberStatDelta(Model):
     membership_no = columns.Text(primary_key=True)
     usage_date = columns.DateTime(primary_key=True)
     time_slice = columns.Text(primary_key=True)
-    previous_cost_per_use = columns.Float()
-    current_cost_per_use = columns.Float()
-    absolute_delta = columns.Integer()
+    previous_cost_per_use = columns.Decimal()
+    current_cost_per_use = columns.Decimal()
+    absolute_delta = columns.Decimal()
     percentage_delta = columns.Float()
 
     __table_name__ = 'subscription_cpu_by_subscriber_stat_delta'
